+++ backend/integration_tests/src/client/user.rs
@@ -21,6 +21,7 @@ generate_update_call!(delete_direct_chat);
 generate_update_call!(delete_group);
 generate_update_call!(delete_messages);
 generate_update_call!(edit_message_v2);
+generate_update_call!(end_video_call);
 generate_update_call!(leave_community);
 generate_update_call!(leave_group);
 generate_update_call!(mark_read);
@@ -31,11 +32,13 @@ generate_update_call!(send_message_v2);
 generate_update_call!(send_message_with_transfer_to_channel);
 generate_update_call!(send_message_with_transfer_to_group);
 generate_update_call!(set_message_reminder_v2);
+generate_update_call!(start_video_call);
 generate_update_call!(tip_message);
 generate_update_call!(unblock_user);
 generate_update_call!(undelete_messages);
 
 pub mod happy_path {
+    use crate::env::VIDEO_CALL_OPERATOR;
     use crate::rng::random_message_id;
     use crate::User;
     use pocket_ic::PocketIc;
@@ -285,5 +288,36 @@ pub mod happy_path {
         );
 
         assert!(matches!(response, user_canister::tip_message::Response::Success))
+    }
+
+    pub fn start_video_call(env: &mut PocketIc, user: &User, recipient: UserId, message_id: MessageId) {
+        let response = super::start_video_call(
+            env,
+            VIDEO_CALL_OPERATOR,
+            recipient.into(),
+            &user_canister::start_video_call::Args {
+                message_id,
+                initiator: user.user_id,
+                initiator_username: user.username(),
+                initiator_display_name: None,
+                initiator_avatar_id: None,
+            },
+        );
+
+        assert!(matches!(response, user_canister::start_video_call::Response::Success(_)))
+    }
+
+    pub fn end_video_call(env: &mut PocketIc, initiator: UserId, recipient: UserId, message_id: MessageId) {
+        let response = super::end_video_call(
+            env,
+            VIDEO_CALL_OPERATOR,
+            recipient.into(),
+            &user_canister::end_video_call::Args {
+                user_id: initiator,
+                message_id,
+            },
+        );
+
+        assert!(matches!(response, user_canister::end_video_call::Response::Success))
     }
 }