+++ backend/libraries/group_chat_core/src/lib.rs
@@ -599,7 +599,6 @@ impl GroupChatCore {
             mentions_disabled,
             everyone_mentioned,
             sender_is_bot,
-            notification_exclusions,
         } = match self.prepare_send_message(
             sender,
             thread_root_message_index,
@@ -692,10 +691,6 @@ impl GroupChatCore {
             }
         }
 
-        for user_id in notification_exclusions {
-            users_to_notify.remove(&user_id);
-        }
-
         Success(SendMessageSuccess {
             message_event,
             users_to_notify: users_to_notify.into_iter().collect(),
@@ -705,7 +700,7 @@ impl GroupChatCore {
 
     fn prepare_send_message(
         &mut self,
-        mut sender: UserId,
+        sender: UserId,
         thread_root_message_index: Option<MessageIndex>,
         content: &MessageContentInternal,
         rules_accepted: Option<Version>,
@@ -714,38 +709,31 @@ impl GroupChatCore {
     ) -> PrepareSendMessageResult {
         use PrepareSendMessageResult::*;
 
-        let mut notification_exclusions = Vec::new();
         if sender == OPENCHAT_BOT_USER_ID || sender == proposals_bot_user_id {
             return Success(PrepareSendMessageSuccess {
                 min_visible_event_index: EventIndex::default(),
                 mentions_disabled: true,
                 everyone_mentioned: false,
                 sender_is_bot: true,
-                notification_exclusions,
             });
         }
 
-        if let MessageContentInternal::VideoCall(vc) = content {
-            sender = vc.participants[0].user_id;
-            notification_exclusions.push(sender);
-        }
-
-        match self.members.get_mut(&sender) {
-            Some(m) => {
-                if m.suspended.value {
-                    return UserSuspended;
-                }
+        if !matches!(content, MessageContentInternal::VideoCall(_)) {
+            if let Some(member) = self.members.get_mut(&sender) {
                 if let Some(version) = rules_accepted {
-                    m.accept_rules(min(version, self.rules.text.version), now);
+                    member.accept_rules(min(version, self.rules.text.version), now);
+                }
+                if !member.check_rules(&self.rules.value) {
+                    return RulesNotAccepted;
                 }
             }
-            None => return UserNotInGroup,
+        }
+
+        let Some(member) = self.members.get(&sender) else {
+            return UserNotInGroup;
         };
-
-        let member = self.members.get(&sender).unwrap();
-
-        if !self.check_rules(member, content) {
-            return RulesNotAccepted;
+        if member.suspended.value {
+            return UserSuspended;
         }
 
         let permissions = &self.permissions;
@@ -762,7 +750,6 @@ impl GroupChatCore {
             mentions_disabled: false,
             everyone_mentioned: member.role.can_mention_everyone(permissions) && is_everyone_mentioned(content),
             sender_is_bot: member.is_bot,
-            notification_exclusions,
         })
     }
 
@@ -1564,16 +1551,6 @@ impl GroupChatCore {
         result
     }
 
-    pub fn check_rules(&self, member: &GroupMemberInternal, content: &MessageContentInternal) -> bool {
-        !self.rules.enabled
-            || member.is_bot
-            || matches!(content, MessageContentInternal::VideoCall(_))
-            || (member
-                .rules_accepted
-                .as_ref()
-                .map_or(false, |accepted| accepted.value >= self.rules.text.version))
-    }
-
     pub fn follow_thread(
         &mut self,
         user_id: UserId,
@@ -2076,5 +2053,4 @@ struct PrepareSendMessageSuccess {
     mentions_disabled: bool,
     everyone_mentioned: bool,
     sender_is_bot: bool,
-    notification_exclusions: Vec<UserId>,
 }